requirements.txt
rembg
Pillow
opencv-python
